chore: updated after review, updated docs
